+++ packages/helpers/src/PromiseQueue.ts
@@ -5,7 +5,7 @@
  *            queue.add(async () => { ... });
  *            await queue.completed;
  */
-export default class PromiseQueue {
+export class PromiseQueue {
   /**
    * Queue of promises to be executed
    */
