+++ packages/helpers/src/deepMerge.ts
@@ -1,0 +1,30 @@
+import { isObject } from "./typeOf";
+
+/**
+ * Merge to objects recursively
+ * @param target - merge target
+ * @param sources - merge sources
+ * @returns
+ */
+export function deepMerge<T extends object>(target, ...sources): T {
+  if (!sources.length) {
+    return target;
+  }
+  const source = sources.shift();
+
+  if (isObject(target) && isObject(source)) {
+    for (const key in source) {
+      if (isObject(source[key])) {
+        if (!target[key]) {
+          Object.assign(target, { [key]: {} });
+        }
+
+        deepMerge(target[key], source[key]);
+      } else {
+        Object.assign(target, { [key]: source[key] });
+      }
+    }
+  }
+
+  return deepMerge(target, ...sources);
+}
