+++ packages/helpers/src/deprecationAssert.ts
@@ -1,0 +1,13 @@
+/**
+ * Common method for printing a warning about the usage of deprecated property or method.
+ * @param condition - condition for deprecation.
+ * @param oldProperty - deprecated property.
+ * @param newProperty - the property that should be used instead.
+ */
+export function deprecationAssert(condition: boolean, oldProperty: string, newProperty: string): void {
+  const message = `«${oldProperty}» is deprecated and will be removed in the next major release. Please use the «${newProperty}» instead.`;
+
+  if (condition) {
+    console.warn(message);
+  }
+}
