+++ packages/helpers/src/keyCodes.ts
@@ -1,0 +1,32 @@
+/**
+ * Returns basic key codes as constants
+ * @returns {{}}
+ */
+export const keyCodes = {
+  BACKSPACE: 8,
+  TAB: 9,
+  ENTER: 13,
+  SHIFT: 16,
+  CTRL: 17,
+  ALT: 18,
+  ESC: 27,
+  SPACE: 32,
+  LEFT: 37,
+  UP: 38,
+  DOWN: 40,
+  RIGHT: 39,
+  DELETE: 46,
+  META: 91,
+  SLASH: 191,
+};
+
+/**
+ * Return mouse buttons codes
+ */
+export const mouseButtons = {
+  LEFT: 0,
+  WHEEL: 1,
+  RIGHT: 2,
+  BACKWARD: 3,
+  FORWARD: 4,
+};
